Fix pre-installation bugs and UI issues. Rev 0.2
Updating the filename removed the .BIN suffix, now we print the whole
filename instead of trying update just the numeric field.

Add a one-second delay before talking to the USB serial port. I don't
know why this is necessary, but it seems to be.

Updated version to 0.2

Reduced the blank space before each prompt, just looks better.

Wait for \n instead of \r, more conventional for terminals and the
Arduino serial monitor.

Output confirmation that we're dumping memory after seizing the bus.

Loop to the right maximum address 0xFFFFF instead of 0x3FFFE.