Document that the USB power jumper must be cut
The Teensy can get power from USB or from the target system
(i.e., the Cadetwriter) but not both. I've chosen to power
the Teensy from the target, on the theory that this is less
stressful for the target's circuits. It does mean that the
Teensy can't be operated outside of a target system without
hooking up some kind of power source, which is inconvenient
for test and debug. You may instead choose to leave the USB
power jumper in place and omit the VDD wire from the DIP clip
to the Teensy, and you'll probably be fine.